requirements.txt
fastapi
uvicorn
langchain
openai
sqlalchemy
psycopg2-binary
python-dotenv
google-generativeai
langchain_community
alembic
